# public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard Management API",
    "version": "1.0.0",
    "description": "A comprehensive API for user authentication and management with role-based permissions",
    "contact": {
      "name": "API Support",
      "email": "app@carik.id"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "01234567-89ab-cdef-0123-456789abcdef",
            "description": "Unique user identifier (UUIDv7)"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "alias": {
            "type": "string",
            "nullable": true,
            "example": "johndoe"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Software Developer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00.000Z"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message description"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Invalid input data"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Invalid or missing authentication token"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Insufficient permissions to access this resource"
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict - Resource already exists",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Resource already exists"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Internal server error"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/v1/group-permissions/{id}": {
      "get": {
        "summary": "Get group permissions",
        "description": "Get group permissions by ID",
        "tags": [
          "Groups"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Client ID for the conversation"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID of the group permission",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group permissions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Group permissions retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "group_permissions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GroupPermission"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/group-permissions": {
      "post": {
        "summary": "Create group permission",
        "description": "Create a new group permission",
        "tags": [
          "Groups"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Client ID for the conversation"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "group_id",
                  "name"
                ],
                "properties": {
                  "group_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the group"
                  },
                  "name": {
                    "type": "string",
                    "description": "Name of the permission"
                  },
                  "resource": {
                    "type": "string",
                    "description": "Resource name"
                  },
                  "action": {
                    "type": "string",
                    "description": "Action name"
                  },
                  "order": {
                    "type": "integer",
                    "description": "Display order"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group permission created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Group permission created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "group_permission": {
                          "$ref": "#/components/schemas/GroupPermission"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/group-permissions/{permissionId}": {
      "put": {
        "summary": "Update group permission",
        "description": "Update an existing group permission",
        "tags": [
          "Groups"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Client ID for the conversation"
          },
          {
            "in": "path",
            "name": "permissionId",
            "required": true,
            "description": "ID of the group permission to update",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the permission"
                  },
                  "resource": {
                    "type": "string",
                    "description": "Resource name"
                  },
                  "action": {
                    "type": "string",
                    "description": "Action name"
                  },
                  "order": {
                    "type": "integer",
                    "description": "Display order"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group permission updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Group permission updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "group_permission": {
                          "$ref": "#/components/schemas/GroupPermission"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete group permission",
        "description": "Delete a group permission (soft delete by setting status_id to 0)",
        "tags": [
          "Groups"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Client ID for the conversation"
          },
          {
            "in": "path",
            "name": "permissionId",
            "required": true,
            "description": "ID of the group permission to delete",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group permission deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Group permission deleted successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/groups": {
      "get": {
        "summary": "List all active groups",
        "description": "Retrieve a list of all active groups (status_id=0) with optional search filtering (requires read permission on groups resource)",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Search term to filter groups by name or description",
            "schema": {
              "type": "string",
              "example": "admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Groups retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Groups retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "groups": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Group"
                          }
                        },
                        "requestedBy": {
                          "$ref": "#/components/schemas/User"
                        },
                        "permissions": {
                          "type": "object",
                          "properties": {
                            "canManage": {
                              "type": "boolean"
                            },
                            "canCreate": {
                              "type": "boolean"
                            },
                            "canEdit": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create new group",
        "description": "Create a new group (requires create permission on groups resource)",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Administrators"
                  },
                  "description": {
                    "type": "string",
                    "example": "System administrators group"
                  },
                  "clientId": {
                    "type": "string",
                    "example": "01234567-89ab-cdef-0123-456789abcdef"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Group created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v1/user/permission": {
      "get": {
        "summary": "Get user permissions",
        "description": "Retrieve the current user's permissions and group memberships",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-client-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Client ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "User permissions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Permission retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userGroupsPermissions": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "groupId": {
                                "type": "integer",
                                "description": "Group ID"
                              },
                              "groupName": {
                                "type": "string",
                                "description": "Group name"
                              },
                              "permissions": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "resource": {
                                      "type": "string",
                                      "description": "Resource name"
                                    },
                                    "action": {
                                      "type": "string",
                                      "description": "Permission action"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication token"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions to access user data"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}